+++ lib/mix/lib/mix/tasks/local.uninstall.ex
@@ -11,7 +11,7 @@ defmodule Mix.Tasks.Local.Uninstall do
   """
 
   def run(argv) do
-    { _, argv } = OptionParser.Simple.parse(argv)
+    { _, argv } = OptionParser.parse(argv)
     Enum.each argv, do_uninstall(&1)
   end